+++ src/print/AutoescapeBlock.js
@@ -1,16 +1,26 @@
 import { doc } from "prettier";
+import { isBoolean } from "lodash";
 import { printChildBlock, quoteChar } from "../util/index.js";
 
 const { hardline } = doc.builders;
+
+const printEscapeType = (escapeType, options) => {
+    if (escapeType === null) {
+        return [];
+    }
+
+    if (isBoolean(escapeType)) {
+        return [escapeType ? "true" : "false", " "];
+    }
+
+    return [quoteChar(options), escapeType, quoteChar(options), " "];
+};
 
 const createOpener = (node, options) => {
     return [
         node.trimLeft ? "{%-" : "{%",
         " autoescape ",
-        quoteChar(options),
-        node.escapeType || "html",
-        quoteChar(options),
-        " ",
+        ...printEscapeType(node.escapeType, options),
         node.trimRightAutoescape ? "-%}" : "%}"
     ];
 };